+++ CHANGELOG.md
@@ -5,16 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.1.6] - 2025-02-28
 
 ### Added
 
 - Grammar to detect document blocks.
+- Beautification of equals to blocks, colons and comments
 
 ### FIXED
 
 - A minor bug fix where config refresh failed due to wrong check.
 - Grammar Syntaxes for data_types.
+
+## [0.1.5] - 2025-02-24
+
+### Changed
+
+- Dropped support for diagnostics completely.
+- Formatter removes blank spaces in empty lines.
 
 ## [0.1.4] - 2025-02-24